double to float conversion didn't compile on Windows
git-svn-id: https://svn.opendtect.org/od/trunk@38716 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/AttributeEngine/attribdataholder.cc
@@ -153,7 +153,7 @@ float DataHolder::getValue( int serieidx, float exactz, float refstep ) const
 
 float DataHolder::getExtraZFromSampPos( float exactz, float refzstep )
 {
-    const float zfactor = SI().zDomain().userFactor();
+    const float zfactor = (float)SI().zDomain().userFactor();
     if ( !(int)(refzstep*zfactor) )
 	return ((exactz/refzstep)-(int)(exactz/refzstep))*refzstep;
